external image adding file
